# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: HypeCheck API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wants:
    get:
      operationId: getWants
      tags: [wants]
      summary: Get the current user's saved products
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Saved wants
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/WantItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "502":
          description: Supabase unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    post:
      operationId: saveWant
      tags: [wants]
      summary: Save a product to Wants
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WantInput"
      responses:
        "201":
          description: Saved want
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WantItem"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "502":
          description: Supabase unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    delete:
      operationId: removeWant
      tags: [wants]
      summary: Remove a product from Wants
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: productId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Want removed
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "502":
          description: Supabase unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"

  # ---------------------------------------------------------------------------
  # Products
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorBody:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string

    # -------------------------------------------------------------------------
    # Auth
    # -------------------------------------------------------------------------
...
    WantItem:
      type: object
      required: [productId, createdAt]
      properties:
        productId:
          type: string
        createdAt:
          type: string
    WantInput:
      type: object
      required: [productId]
      properties:
        productId:
          type: string

    # -------------------------------------------------------------------------
    # Products
    # -------------------------------------------------------------------------
...